+++ docs/coding_style.md
@@ -62,7 +62,7 @@ logger.info("extraneous leading spaces.")
 - Recommend
 
 ```python
-from neural_compressor_ort import logger
+from neural_compressor_ort.utils import logger
 
 logger.info("Current TensorFlow Version is: %s", tf.__version__)  # Use a pattern-string (with %-placeholders)
