+++ neural_compressor_ort/algorithms/layer_wise/core.py
@@ -27,12 +27,11 @@ import onnx
 import onnxruntime as ort
 import transformers
 
-from neural_compressor_ort.common import Logger
+from neural_compressor_ort.utils import logger
 from neural_compressor_ort.quantization.calibrate import CalibrationDataReader
 from neural_compressor_ort.utils.onnx_model import ONNXModel
 from neural_compressor_ort.utils.utility import check_model_with_infer_shapes
 
-logger = Logger().get_logger()
 
 __all__ = [
     "layer_wise_quant",
