+++ neural_compressor_ort/algorithms/smoother/calibrator.py
@@ -24,11 +24,10 @@ import onnx
 import onnx.numpy_helper as numpy_helper
 import onnxruntime
 
-from neural_compressor_ort.common import Logger
+from neural_compressor_ort.utils import logger
 from neural_compressor_ort.quantization.calibrate import CalibrationDataReader
 from neural_compressor_ort.utils.onnx_model import ONNXModel
 
-logger = Logger().get_logger()
 
 __all__ = ["Calibrator"]
