--- neural_compressor_ort/common/__init__.py
@@ -1,40 +1,0 @@
-# Copyright (c) 2023 Intel Corporation
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#    http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-from neural_compressor_ort.common.utils import (
-    level,
-    logger,
-    Logger,
-    TuningLogger,
-    log_quant_execution,
-    set_random_seed,
-    set_resume_from,
-    set_workspace,
-    set_tensorboard,
-    dump_elapsed_time,
-)
-from neural_compressor_ort.common.base_config import options
-
-
-__all__ = [
-    "options",
-    "level",
-    "logger",
-    "Logger",
-    "set_workspace",
-    "set_random_seed",
-    "set_resume_from",
-    "set_tensorboard",
-    "dump_elapsed_time",
-]
