--- neural_compressor_ort/common/utils/__init__.py
@@ -1,20 +1,0 @@
-# Copyright (c) 2023 Intel Corporation
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#    http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-from neural_compressor_ort.common.utils.constants import *
-from neural_compressor_ort.common.utils.logger import *
-from neural_compressor_ort.common.utils.save_load import save_config_mapping, load_config_mapping
-
-# ! Put the following `utility` import after the `logger` import as `utility` used `logger`
-from neural_compressor_ort.common.utils.utility import *
