+++ neural_compressor_ort/quantization/algorithm_entry.py
@@ -20,13 +20,11 @@ import onnx
 from onnxruntime.quantization import quantize
 
 from neural_compressor_ort.algorithms import Smoother
-from neural_compressor_ort.common import Logger
-from neural_compressor_ort.common.utils import AWQ, GPTQ, RTN, SMOOTH_QUANT
+from neural_compressor_ort.utils import AWQ, GPTQ, RTN, SMOOTH_QUANT, logger
 from neural_compressor_ort.quantization.calibrate import CalibrationDataReader
 from neural_compressor_ort.quantization.config import AWQConfig, GPTQConfig, RTNConfig, SmoothQuantConfig
 from neural_compressor_ort.utils.utility import register_algo
 
-logger = Logger().get_logger()
 
 __all__ = [
     "smooth_quant_entry",
