+++ neural_compressor_ort/quantization/matmul_nbits_quantizer.py
@@ -21,8 +21,8 @@ from onnxruntime.quantization.matmul_4bits_quantizer import RTNWeightOnlyQuantCo
 from onnxruntime.quantization.matmul_4bits_quantizer import WeightOnlyQuantConfig
 from packaging import version
 
-from neural_compressor_ort.common import Logger
-from neural_compressor_ort.common.base_config import config_registry
+from neural_compressor_ort.utils import logger
+from neural_compressor_ort.utils.base_config import config_registry
 from neural_compressor_ort.quantization.algorithm_entry import (
     awq_quantize_entry,
     gptq_quantize_entry,
@@ -32,7 +32,6 @@ from neural_compressor_ort.quantization.calibrate import CalibrationDataReader
 from neural_compressor_ort.quantization.config import FRAMEWORK_NAME
 from neural_compressor_ort.utils.onnx_model import ONNXModel
 
-logger = Logger().get_logger()
 
 
 class RTNWeightOnlyQuantConfig(ORTRTNWeightOnlyQuantConfig):
