+++ neural_compressor_ort/quantization/quantize.py
@@ -18,14 +18,11 @@ from typing import Union
 import onnx
 from onnxruntime.quantization.quantize import QuantConfig
 
-from neural_compressor_ort.common import Logger
-from neural_compressor_ort.common.base_config import BaseConfig, ComposableConfig, config_registry
-from neural_compressor_ort.common.utils import log_quant_execution
+from neural_compressor_ort.utils.base_config import BaseConfig, ComposableConfig, config_registry
+from neural_compressor_ort.utils import logger, algos_mapping, log_quant_execution
 from neural_compressor_ort.quantization.calibrate import CalibrationDataReader
 from neural_compressor_ort.quantization.config import FRAMEWORK_NAME
-from neural_compressor_ort.utils.utility import algos_mapping
 
-logger = Logger().get_logger()
 
 
 # ORT-like user-facing API
