+++ neural_compressor_ort/utils/utility.py
@@ -12,28 +12,33 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
+from neural_compressor_ort.utils import logger
 from pathlib import Path
 from typing import Callable, Dict, List, Tuple, Union
-
 import numpy as np
 import onnx
 import onnxruntime.tools.symbolic_shape_infer as symbolic_shape_infer
+import os
 from packaging.version import Version
 
-from neural_compressor_ort.common import Logger
+import importlib
+import subprocess
+import time
 
-logger = Logger().get_logger()
+import cpuinfo
+import psutil
+
 
 __all__ = [
-    "ONNXRT116_VERSION",
-    "ONNXRT1161_VERSION",
+    "set_workspace",
+    "set_random_seed",
+    "set_resume_from",
+    "dump_elapsed_time",
+    "log_quant_execution",
+    "singleton",
+    "LazyImport",
+    "CpuInfo",
     "algos_mapping",
-    "WHITE_MODULE_LIST",
-    "MAXIMUM_PROTOBUF",
-    "PRIORITY_RTN",
-    "PRIORITY_GPTQ",
-    "PRIORITY_AWQ",
-    "PRIORITY_SMOOTH_QUANT",
     "dtype_mapping",
     "find_by_name",
     "simple_progress_bar",
@@ -45,21 +50,184 @@ __all__ = [
     "check_model_with_infer_shapes",
 ]
 
-ONNXRT116_VERSION = Version("1.16.0")
-ONNXRT1161_VERSION = Version("1.16.1")
 
 # Dictionary to store a mapping between algorithm names and corresponding algo implementation(function)
 algos_mapping: Dict[str, Callable] = {}
 
-# All constants for onnxrt
-WHITE_MODULE_LIST = ["MatMul", "Conv"]
 
-MAXIMUM_PROTOBUF = 2147483648
+def singleton(cls):
+    """Singleton decorator."""
 
-PRIORITY_RTN = 60
-PRIORITY_GPTQ = 70
-PRIORITY_AWQ = 50
-PRIORITY_SMOOTH_QUANT = 80
+    instances = {}
+
+    def _singleton(*args, **kw):
+        """Create a singleton object."""
+        if cls not in instances:
+            instances[cls] = cls(*args, **kw)
+        return instances[cls]
+
+    return _singleton
+
+
+class LazyImport(object):
+    """Lazy import python module till use."""
+
+    def __init__(self, module_name):
+        """Init LazyImport object.
+
+        Args:
+           module_name (string): The name of module imported later
+        """
+        self.module_name = module_name
+        self.module = None
+
+    def __getattr__(self, name):
+        """Get the attributes of the module by name."""
+        try:
+            self.module = importlib.import_module(self.module_name)
+            mod = getattr(self.module, name)
+        except:
+            spec = importlib.util.find_spec(str(self.module_name + "." + name))
+            mod = importlib.util.module_from_spec(spec)
+            spec.loader.exec_module(mod)
+        return mod
+
+    def __call__(self, *args, **kwargs):
+        """Call the function in that module."""
+        function_name = self.module_name.split(".")[-1]
+        module_name = self.module_name.split(f".{function_name}")[0]
+        self.module = importlib.import_module(module_name)
+        function = getattr(self.module, function_name)
+        return function(*args, **kwargs)
+
+
+@singleton
+class CpuInfo(object):
+    """CPU info collection."""
+
+    def __init__(self):
+        """Get whether the cpu numerical format is bf16, the number of sockets, cores and cores per socket."""
+        self._bf16 = False
+        self._vnni = False
+        info = cpuinfo.get_cpu_info()
+        if "arch" in info and "X86" in info["arch"]:
+            cpuid = cpuinfo.CPUID()
+            max_extension_support = cpuid.get_max_extension_support()
+            if max_extension_support >= 7:
+                ecx = cpuid._run_asm(
+                    b"\x31\xC9",  # xor ecx, ecx
+                    b"\xB8\x07\x00\x00\x00" b"\x0f\xa2" b"\x89\xC8" b"\xC3",  # mov eax, 7  # cpuid  # mov ax, cx  # ret
+                )
+                self._vnni = bool(ecx & (1 << 11))
+                eax = cpuid._run_asm(
+                    b"\xB9\x01\x00\x00\x00",  # mov ecx, 1
+                    b"\xB8\x07\x00\x00\x00" b"\x0f\xa2" b"\xC3",  # mov eax, 7  # cpuid  # ret
+                )
+                self._bf16 = bool(eax & (1 << 5))
+        # TODO: The implementation will be refined in the future.
+        # https://github.com/intel/neural-compressor/tree/detect_sockets
+        if "arch" in info and "ARM" in info["arch"]:  # pragma: no cover
+            self._sockets = 1
+        else:
+            self._sockets = self.get_number_of_sockets()
+        self._cores = psutil.cpu_count(logical=False)
+        self._cores_per_socket = int(self._cores / self._sockets)
+
+    @property
+    def bf16(self):
+        """Get whether it is bf16."""
+        return self._bf16
+
+    @property
+    def vnni(self):
+        """Get whether it is vnni."""
+        return self._vnni
+
+    @property
+    def cores_per_socket(self):
+        """Get the cores per socket."""
+        return self._cores_per_socket
+
+    def get_number_of_sockets(self) -> int:
+        """Get number of sockets in platform."""
+        cmd = "cat /proc/cpuinfo | grep 'physical id' | sort -u | wc -l"
+        if psutil.WINDOWS:
+            cmd = r'wmic cpu get DeviceID | C:\Windows\System32\find.exe /C "CPU"'
+        elif psutil.MACOS:  # pragma: no cover
+            cmd = "sysctl -n machdep.cpu.core_count"
+
+        with subprocess.Popen(
+            args=cmd,
+            shell=True,
+            stdout=subprocess.PIPE,
+            stderr=subprocess.STDOUT,
+            universal_newlines=False,
+        ) as proc:
+            proc.wait()
+            if proc.stdout:
+                for line in proc.stdout:
+                    return int(line.decode("utf-8", errors="ignore").strip())
+        return 0
+
+
+def dump_elapsed_time(customized_msg=""):
+    """Get the elapsed time for decorated functions.
+
+    Args:
+        customized_msg (string, optional): The parameter passed to decorator. Defaults to None.
+    """
+
+    def f(func):
+        def fi(*args, **kwargs):
+            start = time.time()
+            res = func(*args, **kwargs)
+            end = time.time()
+            logger.info(
+                "%s elapsed time: %s ms"
+                % (customized_msg if customized_msg else func.__qualname__, round((end - start) * 1000, 2))
+            )
+            return res
+
+        return fi
+
+    return f
+
+
+def set_random_seed(seed: int):
+    """Set the random seed in config."""
+    from neural_compressor_ort.utils.base_config import options
+
+    options.random_seed = seed
+
+
+def set_workspace(workspace: str):
+    """Set the workspace in config."""
+    from neural_compressor_ort.utils.base_config import options
+
+    options.workspace = workspace
+
+
+def set_resume_from(resume_from: str):
+    """Set the resume_from in config."""
+    from neural_compressor_ort.utils.base_config import options
+
+    options.resume_from = resume_from
+
+
+def log_quant_execution(func):
+    from neural_compressor_ort.utils import TuningLogger
+    default_tuning_logger = TuningLogger()
+    def wrapper(*args, **kwargs):
+        default_tuning_logger.quantization_start(stacklevel=4)
+
+        # Call the original function
+        result = func(*args, **kwargs)
+
+        default_tuning_logger.quantization_end(stacklevel=4)
+        return result
+
+    return wrapper
+
 
 dtype_mapping = {
     "fp32": 1,
