+++ test/quantization/weight_only/test_gptq.py
@@ -3,11 +3,12 @@ import os
 import shutil
 import unittest
 
+import torch
 from optimum.exporters.onnx import main_export
+from transformers import AutoTokenizer
 
-from neural_compressor_ort.common import Logger
-
-logger = Logger().get_logger()
+from neural_compressor_ort.utils import logger
+from neural_compressor_ort.quantization.calibrate import CalibrationDataReader
 
 
 def find_onnx_file(folder_path):
@@ -19,7 +20,35 @@ def find_onnx_file(folder_path):
     return None
 
 
-class TestRTNQuant(unittest.TestCase):
+class DummyNLPDataloader(CalibrationDataReader):
+    def __init__(self, model_name):
+        self.tokenizer = AutoTokenizer.from_pretrained(model_name)
+        self.sequence_a = "intel-extension-for-transformers is based in SH"
+        self.sequence_b = "Where is intel-extension-for-transformers based? NYC or SH"
+
+        self.encoded_list = []
+        encoded_input = dict(self.tokenizer(self.sequence_a, self.sequence_b, return_tensors="pt"))
+        input_shape = encoded_input["input_ids"].shape
+        encoded_input["position_ids"] = (
+            torch.arange(0, input_shape[-1], dtype=torch.long).unsqueeze(0).view(-1, input_shape[-1])
+        )
+
+        # convert torch tensor to numpy
+        for input_name, input_value in encoded_input.items():
+            if isinstance(input_value, torch.Tensor):
+                encoded_input[input_name] = input_value.numpy()
+
+        self.encoded_list.append(encoded_input)
+        self.iter_next = iter(self.encoded_list)
+
+    def get_next(self):
+        return next(self.iter_next, None)
+
+    def rewind(self):
+        self.iter_next = iter(self.encoded_list)
+
+
+class TestGPTQQuant(unittest.TestCase):
     @classmethod
     def setUpClass(self):
         main_export(
@@ -27,6 +56,7 @@ class TestRTNQuant(unittest.TestCase):
             output="gptj",
         )
         self.gptj = find_onnx_file("./gptj")
+        self.calibration_data_reader = DummyNLPDataloader("hf-internal-testing/tiny-random-gptj")
 
     @classmethod
     def tearDownClass(self):
@@ -34,7 +64,15 @@ class TestRTNQuant(unittest.TestCase):
 
     def setUp(self):
         # print the test name
-        logger.info(f"Running ONNXRT TestRTNQuant test: {self.id()}")
+        logger.info(f"Running ONNXRT TestGPTQQuant test: {self.id()}")
+
+    def _count_woq_matmul(self, q_model, bits=4, group_size=32):
+        op_names = [
+            i.name
+            for i in q_model.graph.node
+            if i.op_type.startswith("MatMul") and i.input[1].endswith("_Q{}G{}".format(bits, group_size))
+        ]
+        return len(op_names)
 
     def _check_model_is_quantized(self, model):
         node_optypes = [node.op_type for node in model.graph.node]
@@ -49,109 +87,101 @@ class TestRTNQuant(unittest.TestCase):
                 return True
         return False
 
-    def _count_woq_matmul(self, q_model, bits=4, group_size=32):
-        op_names = [
-            i.name
-            for i in q_model.graph.node
-            if i.op_type.startswith("MatMul") and i.input[1].endswith("_Q{}G{}".format(bits, group_size))
-        ]
-        return len(op_names)
-
-    def _apply_rtn(self, quant_config):
-        logger.info(f"Test RTN with config {quant_config}")
+    def _apply_gptq(self, quant_config):
+        logger.info(f"Test GPTQ with config {quant_config}")
         from neural_compressor_ort.quantization.quantize import _quantize
 
         fp32_model = copy.deepcopy(self.gptj)
-        qmodel = _quantize(fp32_model, quant_config)
+        qmodel = _quantize(fp32_model, quant_config, calibration_data_reader=self.calibration_data_reader)
         self.assertIsNotNone(qmodel)
         return qmodel
 
 
-class TestRTNQuantWithInternalAPI(TestRTNQuant):
-    def test_rtn_params_combination(self):
-        from neural_compressor_ort.quantization import RTNConfig
+class TestGPTQQuantWithInternalAPI(TestGPTQQuant):
+    def test_gptq_params_combination(self):
+        from neural_compressor_ort.quantization import GPTQConfig
 
         # some tests were skipped to accelerate the CI
         # TODO: check params combination.
         # TODO: Add number check for group_size.
-        rtn_options = {
+        gptq_options = {
             "weight_dtype": ["int"],
-            "weight_bits": [4, 3, 8],
+            "weight_bits": [4],
             "weight_group_size": [32],
             "weight_sym": [True, False],
             "act_dtype": ["fp32"],
+            "accuracy_level": [0],
+            "percdamp": [0.01],
+            "blocksize": [128],
+            "actorder": [True, False],
+            "mse": [True, False],
+            "perchannel": [True, False],
         }
         from itertools import product
 
-        keys = RTNConfig.params_list
-        for value in product(*rtn_options.values()):
+        keys = GPTQConfig.params_list
+        for value in product(*gptq_options.values()):
             d = dict(zip(keys, value))
-            quant_config = RTNConfig(**d)
-            qmodel = self._apply_rtn(quant_config)
+            print(d)
+            quant_config = GPTQConfig(**d)
+            qmodel = self._apply_gptq(quant_config)
             self.assertEqual(self._count_woq_matmul(qmodel, bits=value[1], group_size=value[2]), 30)
 
-    def test_rtn_config(self):
-        from neural_compressor_ort.quantization import RTNConfig
+    def test_gptq_config(self):
+        from neural_compressor_ort.quantization import GPTQConfig
 
-        rtn_config1 = RTNConfig(weight_bits=4)
+        gptq_config1 = GPTQConfig(weight_bits=4)
         quant_config_dict = {
-            "rtn": {"weight_bits": 4},
+            "gptq": {"weight_bits": 4},
         }
-        rtn_config2 = RTNConfig.from_dict(quant_config_dict["rtn"])
-        self.assertEqual(rtn_config1.to_dict(), rtn_config2.to_dict())
+        gptq_config2 = GPTQConfig.from_dict(quant_config_dict["gptq"])
+        self.assertEqual(gptq_config1.to_dict(), gptq_config2.to_dict())
 
-    def test_quantize_rtn_from_dict_default(self):
-        from neural_compressor_ort.quantization import get_default_rtn_config
-        from neural_compressor_ort.quantization.quantize import _quantize
+    def test_quantize_gptq_from_dict_default(self):
+        from neural_compressor_ort.quantization import get_default_gptq_config
 
-        qmodel = self._apply_rtn(quant_config=get_default_rtn_config())
+        qmodel = self._apply_gptq(quant_config=get_default_gptq_config())
         self.assertIsNotNone(qmodel)
         self.assertTrue(self._check_model_is_quantized(qmodel))
 
-    def test_quantize_rtn_from_dict_beginner(self):
-        from neural_compressor_ort.quantization.quantize import _quantize
-
+    def test_quantize_gptq_from_dict_beginner(self):
         quant_config = {
-            "rtn": {
+            "gptq": {
                 "weight_bits": 4,
                 "weight_group_size": 32,
             },
         }
-        qmodel = self._apply_rtn(quant_config)
+        qmodel = self._apply_gptq(quant_config)
         self.assertIsNotNone(qmodel)
         self.assertIsNotNone(qmodel)
         self.assertTrue(self._check_model_is_quantized(qmodel))
 
-    def test_quantize_rtn_from_class_beginner(self):
-        from neural_compressor_ort.quantization import RTNConfig
-        from neural_compressor_ort.quantization.quantize import _quantize
+    def test_quantize_gptq_from_class_beginner(self):
+        from neural_compressor_ort.quantization import GPTQConfig
 
-        quant_config = RTNConfig(weight_bits=4, weight_group_size=32)
-        qmodel = self._apply_rtn(quant_config)
+        quant_config = GPTQConfig(weight_bits=4, weight_group_size=32)
+        qmodel = self._apply_gptq(quant_config)
         self.assertIsNotNone(qmodel)
 
-    def test_quantize_rtn_fallback_from_class_beginner(self):
-        from neural_compressor_ort.quantization import RTNConfig
-        from neural_compressor_ort.quantization.quantize import _quantize
+    def test_quantize_gptq_fallback_from_class_beginner(self):
+        from neural_compressor_ort.quantization import GPTQConfig
 
-        fp32_config = RTNConfig(weight_dtype="fp32")
-        quant_config = RTNConfig(
+        fp32_config = GPTQConfig(weight_dtype="fp32")
+        quant_config = GPTQConfig(
             weight_bits=4,
             weight_dtype="int",
             weight_sym=False,
             weight_group_size=32,
         )
         quant_config.set_local("/h.4/mlp/fc_out/MatMul", fp32_config)
-        qmodel = self._apply_rtn(quant_config)
+        qmodel = self._apply_gptq(quant_config)
         self.assertIsNotNone(qmodel)
         self.assertEqual(self._count_woq_matmul(qmodel), 29)
         self.assertFalse(self._check_node_is_quantized(qmodel, "/h.4/mlp/fc_out/MatMul"))
 
-    def test_quantize_rtn_from_dict_advance(self):
-        from neural_compressor_ort.quantization.quantize import _quantize
-
+    def test_quantize_gptq_from_dict_advance(self):
         quant_config = {
-            "rtn": {
+            "gptq": {
                 "global": {
                     "weight_bits": 4,
                     "weight_group_size": 32,
@@ -163,13 +193,13 @@ class TestRTNQuantWithInternalAPI(TestRTNQuant):
                 },
             }
         }
-        qmodel = self._apply_rtn(quant_config)
+        qmodel = self._apply_gptq(quant_config)
         self.assertIsNotNone(qmodel)
         self.assertEqual(self._count_woq_matmul(qmodel), 29)
         self.assertFalse(self._check_node_is_quantized(qmodel, "/h.4/mlp/fc_out/MatMul"))
 
         quant_config = {
-            "rtn": {
+            "gptq": {
                 "global": {
                     "weight_bits": 4,
                     "weight_group_size": 32,
@@ -182,18 +212,20 @@ class TestRTNQuantWithInternalAPI(TestRTNQuant):
                 },
             }
         }
-        qmodel = self._apply_rtn(quant_config)
+        qmodel = self._apply_gptq(quant_config)
         self.assertIsNotNone(qmodel)
         for node in qmodel.graph.node:
             if node.name == "/h.4/mlp/fc_out/MatMul":
                 self.assertTrue(node.input[1].endswith("Q8G32"))
 
 
-class TestRTNQuantWithORTLikeAPI(TestRTNQuant):
-    def test_rtn_config_4bits(self):
+class TestGPTQQuantWithORTLikeAPI(TestGPTQQuant):
+    def test_gptq_config_4bits(self):
         from neural_compressor_ort.quantization import matmul_4bits_quantizer
 
-        algo_config = matmul_4bits_quantizer.RTNWeightOnlyQuantConfig()
+        algo_config = matmul_4bits_quantizer.GPTQWeightOnlyQuantConfig(
+            calibration_data_reader=self.calibration_data_reader
+        )
 
         quant = matmul_4bits_quantizer.MatMul4BitsQuantizer(
             copy.deepcopy(self.gptj),
@@ -205,10 +237,12 @@ class TestRTNQuantWithORTLikeAPI(TestRTNQuant):
         self.assertIsNotNone(quant.model)
         self.assertTrue(self._check_model_is_quantized(quant.model))
 
-    def test_rtn_config_4bits_with_exclude_node(self):
+    def test_gptq_config_4bits_with_exclude_node(self):
         from neural_compressor_ort.quantization import matmul_4bits_quantizer
 
-        algo_config = matmul_4bits_quantizer.RTNWeightOnlyQuantConfig()
+        algo_config = matmul_4bits_quantizer.GPTQWeightOnlyQuantConfig(
+            calibration_data_reader=self.calibration_data_reader
+        )
 
         quant = matmul_4bits_quantizer.MatMul4BitsQuantizer(
             copy.deepcopy(self.gptj),
@@ -222,10 +256,12 @@ class TestRTNQuantWithORTLikeAPI(TestRTNQuant):
         self.assertTrue(self._check_model_is_quantized(quant.model))
         self.assertFalse(self._check_node_is_quantized(quant.model, "/h.4/mlp/fc_out/MatMul"))
 
-    def test_rtn_config_nbits(self):
+    def test_gptq_config_nbits(self):
         from neural_compressor_ort.quantization import matmul_nbits_quantizer
 
-        algo_config = matmul_nbits_quantizer.RTNWeightOnlyQuantConfig()
+        algo_config = matmul_nbits_quantizer.GPTQWeightOnlyQuantConfig(
+            calibration_data_reader=self.calibration_data_reader
+        )
 
         for n_bits in [3, 4, 8]:
             quant = matmul_nbits_quantizer.MatMulNBitsQuantizer(
@@ -239,10 +275,12 @@ class TestRTNQuantWithORTLikeAPI(TestRTNQuant):
             self.assertIsNotNone(quant.model)
             self.assertEqual(self._count_woq_matmul(quant.model, bits=n_bits, group_size=32), 30)
 
-    def test_rtn_config_nbits_with_exclude_node(self):
+    def test_gptq_config_nbits_with_exclude_node(self):
         from neural_compressor_ort.quantization import matmul_nbits_quantizer
 
-        algo_config = matmul_nbits_quantizer.RTNWeightOnlyQuantConfig()
+        algo_config = matmul_nbits_quantizer.GPTQWeightOnlyQuantConfig(
+            calibration_data_reader=self.calibration_data_reader
+        )
 
         for n_bits in [3, 4, 8]:
             quant = matmul_nbits_quantizer.MatMulNBitsQuantizer(
