+++ test/utils/test_general.py
@@ -1,57 +1,20 @@
-"""Tests for common components.
-
-!!! Please do not import any framework-specific modules in this file. !!!
-* Note, we may need to add some auto check mechanisms to ensure this.
-
-These tests aim to assess the fundamental functionalities of common components and enhance code coverage.
-All tests will be included for each framework CI.
-
-* Note
-The folder structure:
-.
-├── 3x
-│   ├── common
-│   ├── onnxrt
-│   ├── tensorflow
-│   └── torch
-
-For each fwk CI:
-
-onnxrt_included_folder:
-    ├── 3x
-    │   ├── common
-    │   ├── onnxrt
-
-tensorflow_included_folder:
-    ├── 3x
-    │   ├── common
-    │   ├── tensorflow
-
-
-torch_included_folder:
-    ├── 3x
-    │   ├── common
-    │   ├── torch
-"""
+"""Tests for utils components."""
 
 import unittest
 
-from neural_compressor_ort.common import Logger
-
-logger = Logger().get_logger()
-
 from typing import Any, List, Optional, Tuple, Union
 
-from neural_compressor_ort.common.base_config import (
+from neural_compressor_ort.utils.base_config import (
     BaseConfig,
     config_registry,
     get_all_config_set_from_config_registry,
     register_config,
     register_supported_configs_for_fwk,
 )
-from neural_compressor_ort.common.base_tuning import ConfigLoader, ConfigSet, Evaluator, SequentialSampler
-from neural_compressor_ort.common.tuning_param import TuningParam
-from neural_compressor_ort.common.utils import DEFAULT_WHITE_LIST, OP_NAME_OR_MODULE_TYPE
+from neural_compressor_ort.utils.tuning_param import TuningParam
+from neural_compressor_ort.utils.base_tuning import ConfigLoader, ConfigSet, Evaluator, SequentialSampler
+from neural_compressor_ort.utils import logger
+from neural_compressor_ort.utils.constants import DEFAULT_WHITE_LIST, OP_NAME_OR_MODULE_TYPE
 
 PRIORITY_FAKE_ALGO = 100
 FAKE_CONFIG_NAME = "fake"
