+++ test/utils/test_param.py
@@ -1,7 +1,9 @@
+"""Tests for utils components."""
+
 import unittest
 from typing import List
 
-from neural_compressor_ort.common.tuning_param import TuningParam
+from neural_compressor_ort.utils.tuning_param import TuningParam
 
 
 class TestTuningParam(unittest.TestCase):
